Feat: Adds gpio checks for HIGH signals
Summary:
-----
In order to support external sensors and microcontrollers, I added
reads of a list of 3 GPIO pins. If any of those is HIGH, then the
schokotron should open the door. This should enable non twitter
opening mechanisms.

Motivation:
-----
Currently we are planning to have non twitter access mechnisms, which
could be either a sensor that outputs a HIGH signal or a dedicated
microcontroller like an arduino or feather. They can be programmed
independently and should allow us adding new features with low-overhead.